Start trying to get texconv working with FastBuild.
The main issue is that texconv expects the output to be a folder not a file name, but fastbuild expects an exact file name.
I think the easiest way to solve this is to modify texconv to take a new commandline arg which is a target file name, rather than a folder.